fix(sherlock-rxjs): fix possible stale-cache / memory leak in fromObservable
BREAKING CHANGE: This changes the behavior of `fromObservable` when not connected. From now on, a Derivable created by `fromObservable` will become unresolved when disconnected. In previous versions the Derivable would remember the last received value which can lead to unexpected stale results when reconnecting or possibly memory leaks. To get the old behavior use `take`: `fromObservable(...).take({ when: d => d.resolved })` to always cache the latest value when disconnecting.