feat(sherlock): modernize types
With new TypeScript features available to us now, we can remove most of the overloaded methods from the interfaces, so we now support an arbitrary number of parameters to `derive`, `#swap` and `lens`. `#flatMap` is now also synchronised with `Array#flatMap` so you can either return a Derivable or an arbitrary value.